requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
pandas
scikit-learn
numpy
regex
requests
streamlit
python-dotenv
pytest
